.ares/applications/sales-coach/requirements.txt
# Sales Coach - Real-Time System Requirements

# Core AI/ML
anthropic>=0.40.0
deepgram-sdk>=3.0.0

# Audio Processing
pyaudio>=0.2.13
sounddevice>=0.4.6
numpy>=1.24.0

# Web Server & UI
flask>=3.0.0
flask-cors>=4.0.0
flask-socketio>=5.3.0
python-socketio>=5.10.0

# Data & Config
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
websockets>=12.0
aiohttp>=3.9.0
